# apps/post-store/src/server/go/model_ai_content_post_request.go
/*
 * Post Store API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type AiContentPostRequest struct {
	Text string `json:"text"`

	Actor string `json:"actor"`
}

// AssertAiContentPostRequestRequired checks if the required fields are not zero-ed
func AssertAiContentPostRequestRequired(obj AiContentPostRequest) error {
	elements := map[string]interface{}{
		"text":  obj.Text,
		"actor": obj.Actor,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertRecurseAiContentPostRequestRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of AiContentPostRequest (e.g. [][]AiContentPostRequest), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseAiContentPostRequestRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aAiContentPostRequest, ok := obj.(AiContentPostRequest)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertAiContentPostRequestRequired(aAiContentPostRequest)
	})
}