# apps/post-store/src/server/go/model_ai_content_response.go
/*
 * Post Store API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type AiContentResponse struct {
	Text string `json:"text"`

	RateLimit AiContentResponseRateLimit `json:"rateLimit"`
}

// AssertAiContentResponseRequired checks if the required fields are not zero-ed
func AssertAiContentResponseRequired(obj AiContentResponse) error {
	elements := map[string]interface{}{
		"text":      obj.Text,
		"rateLimit": obj.RateLimit,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	if err := AssertAiContentResponseRateLimitRequired(obj.RateLimit); err != nil {
		return err
	}
	return nil
}

// AssertRecurseAiContentResponseRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of AiContentResponse (e.g. [][]AiContentResponse), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseAiContentResponseRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aAiContentResponse, ok := obj.(AiContentResponse)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertAiContentResponseRequired(aAiContentResponse)
	})
}